sokogate-subscription-model/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations

# Pydantic for data validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Payment processing
stripe==7.8.0

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Configuration management
python-decouple==3.8
pyyaml==6.0.1

# HTTP client for external API calls
httpx==0.25.2
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# Email and notifications
sendgrid==6.11.0
twilio==8.11.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI endpoints
faker==20.1.0  # For generating test data

# Development tools
black==23.11.0  # Code formatting
flake8==6.1.0  # Code linting
mypy==1.7.1  # Type checking
pre-commit==3.6.0  # Git hooks

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Environment and deployment
python-dotenv==1.0.0
gunicorn==21.2.0
docker==6.1.3

# Analytics and metrics
prometheus-client==0.19.0
statsd==4.0.1

# Utilities
click==8.1.7  # CLI utilities
jinja2==3.1.2  # Template engine
python-slugify==8.0.1
cryptography==41.0.8

# API documentation
fastapi-users==12.1.2
fastapi-pagination==0.12.13

# Webhook handling
webhooks==0.0.11

# Data processing
pandas==2.1.3
numpy==1.25.2

# File handling
openpyxl==3.1.2  # Excel file handling
Pillow==10.1.0  # Image processing
